refactor(dh): remove related metering points v1
